RT#99458: use pkgconfig to detect libhunspell
This is the correct way to figure out where libhunspell is,
and I believe the one that will work on most systems that use
pkgconfig at least. The previous method was not very generic,
to say the least.

Thanks to David Gutteridge for providing a patch, showing me
how to do this. Let's see what feedback we get from CPAN
testers.

https://rt.cpan.org/Ticket/Display.html?id=99548